refactor(db): extract DB operations into shared @bahar/db-operations package
Move all local-DB operation logic (decks, dictionary-entries, flashcards,
settings, progress, migration) out of apps/web into @bahar/db-operations as
factory functions: makeXTable({ getDb }) => { ...operations }. Each app injects
its own DB accessor via getDb; the operation logic is now platform-agnostic and
shared, ready for mobile to consume the same code.

Key pieces:
- Operations are drizzle-only. The two former raw-SQL holdouts are handled:
  flashcards.counts is now expressed in drizzle (mirroring decks.list); the
  dictionary-entries.tags json_each aggregate stays raw via drizzleDb.all(sql)
  with a positional map, since json_each in FROM has no query-builder form.
- The write queue moves into the package (queue.ts): enqueueDbOperation plus
  the sync-merge helpers (enqueueSyncOperation, hasPendingOperations). It has no
  logging dependency of its own -- apps wire their Sentry logger once via
  configureDbQueue({ onError, onInfo }). Web does this in lib/db/index.ts, the
  universal init chokepoint, so the handlers are set before any queued op runs.
- clearBacklog stays engine-agnostic in the package; its web-specific revlog
  POST + Sentry reporting is injected as a flashcards-factory second argument
  (postRevlogBatch), keeping the shared OperationDeps at just { getDb }.
- New real-DB test harness (src/test/): @tursodatabase/sync runs the same Turso
  engine family as web (sync-wasm) and mobile (sync-react-native) but in plain
  Node, so operations are tested against a real in-memory SQLite DB via vitest
  -- no mocks. 93 tests cover every operation, including the same-name-column
  join-collision the aliasing convention guards against.

Web is now pure wiring: apps/web/src/lib/db/operations.ts instantiates the
factories with web's DB singleton, and its only remaining tests exercise
web-specific surface (the sync-wasm adapter smoke test and clearBacklog's API
wiring). Consumers import shared types/constants/queue fns straight from
@bahar/db-operations rather than through a web re-export.

Adds date-fns, @bahar/fsrs, ts-fsrs, @tanstack/pacer, and (dev)
@tursodatabase/sync + vitest to the package.
