+++ apps/web/src/components/features/dictionary/filters/TagsFilter.tsx
@@ -19,7 +19,7 @@ import { useQuery } from "@tanstack/react-query";
 import { useNavigate, useSearch } from "@tanstack/react-router";
 import { CheckIcon, ChevronsUpDownIcon } from "lucide-react";
 import { useState } from "react";
-import { dictionaryEntriesTable } from "@/lib/db/operations/dictionary-entries";
+import { dictionaryEntriesTable } from "@/lib/db/operations";
 
 export const TagsFilter = () => {
   const [open, setOpen] = useState(false);
