+++ apps/web/src/components/features/flashcards/FlashcardDrawer/TagBadgesList.tsx
@@ -1,9 +1,9 @@
+import type { FlashcardWithDictionaryEntry } from "@bahar/db-operations";
 import type { SelectDictionaryEntry } from "@bahar/drizzle-user-db-schemas";
 import { Badge } from "@bahar/web-ui/components/badge";
 import { t } from "@lingui/core/macro";
 import { motion } from "motion/react";
 import type { FC } from "react";
-import type { FlashcardWithDictionaryEntry } from "@/lib/db/operations/flashcards";
 
 const getTranslatedType = (entryType: SelectDictionaryEntry["type"]) => {
   switch (entryType) {
