+++ apps/web/src/hooks/db/index.ts
@@ -3,8 +3,7 @@ import { useMutation } from "@tanstack/react-query";
 import { useSetAtom } from "jotai";
 import { createEmptyCard } from "ts-fsrs";
 import { suggestedTagsAtom } from "@/atoms/suggested-tags";
-import { dictionaryEntriesTable } from "@/lib/db/operations/dictionary-entries";
-import { flashcardsTable } from "@/lib/db/operations/flashcards";
+import { dictionaryEntriesTable, flashcardsTable } from "@/lib/db/operations";
 import { queryClient } from "@/lib/query";
 import { getOramaDb, toOramaDocument } from "@/lib/search";
 import { useSearch } from "../search/useSearch";
