+++ apps/web/src/routes/_authorized-layout/_app-layout/dictionary/edit/$wordId.tsx
@@ -49,8 +49,7 @@ import { useDeleteDictionaryEntry, useEditDictionaryEntry } from "@/hooks/db";
 import { useDir } from "@/hooks/useDir";
 import { useUserPlan } from "@/hooks/useUserPlan";
 import { api } from "@/lib/api";
-import { dictionaryEntriesTable } from "@/lib/db/operations/dictionary-entries";
-import { flashcardsTable } from "@/lib/db/operations/flashcards";
+import { dictionaryEntriesTable, flashcardsTable } from "@/lib/db/operations";
 import { queryClient } from "@/lib/query";
 import { FormSchema } from "@/lib/schemas/dictionary";
 import type { z } from "@/lib/zod";
