+++ packages/db-operations/src/operations/decks.test.ts
@@ -3,31 +3,27 @@ import {
   FlashcardState,
   type SelectDeck,
 } from "@bahar/drizzle-user-db-schemas";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { createTestDb, type TestDb } from "../test/create-test-db";
 import {
   insertDeck,
   insertDictionaryEntry,
   insertFlashcard,
 } from "../test/factories";
-
-const dbRef = vi.hoisted(() => ({ current: undefined as TestDb | undefined }));
-
-vi.mock("..", async (importOriginal) => ({
-  ...(await importOriginal()),
-  ensureDb: vi.fn(async () => dbRef.current?.db),
-  getDb: vi.fn(() => dbRef.current?.db),
-  getDrizzleDb: vi.fn(() => dbRef.current?.drizzleDb),
-}));
-
-const { decksTable } = await import("./decks");
+import { makeDecksTable } from "./decks";
 
 describe("decksTable", () => {
   let testDb: TestDb;
+  let decksTable: ReturnType<typeof makeDecksTable>;
 
   beforeEach(async () => {
     testDb = await createTestDb();
-    dbRef.current = testDb;
+    // Inject the real harness db directly -- no mocks. Mutations run through
+    // the package's real write queue (imported by the factory), so the tests
+    // exercise the actual serialized path.
+    decksTable = makeDecksTable({
+      getDb: async () => testDb.drizzleDb,
+    });
   });
 
   afterEach(async () => {
@@ -45,9 +41,7 @@ describe("decksTable", () => {
         },
       } satisfies Omit<SelectDeck, "id">;
 
-      const newDeck = await decksTable.create.mutation({
-        deck: deckInput,
-      });
+      const newDeck = await decksTable.create.mutation({ deck: deckInput });
 
       expect(newDeck).toMatchObject({
         id: expect.any(String),
@@ -59,26 +53,23 @@ describe("decksTable", () => {
         },
       });
 
-      const deckExists = await testDb.db
-        .prepare("SELECT * FROM decks WHERE id = ?")
-        .get([newDeck.id]);
+      const deckExists = await (
+        await testDb.db.prepare("SELECT * FROM decks WHERE id = ?")
+      ).get([newDeck.id]);
 
       expect(deckExists).toBeDefined();
     });
 
     it("stores null filters as null, not the string 'null'", async () => {
       const newDeck = await decksTable.create.mutation({
-        deck: {
-          name: "test deck",
-          filters: null,
-        },
+        deck: { name: "test deck", filters: null },
       });
 
       expect(newDeck.filters).toBeNull();
 
-      const deckExists = await testDb.db
-        .prepare("SELECT * FROM decks WHERE id = ?")
-        .get([newDeck.id]);
+      const deckExists = (await (
+        await testDb.db.prepare("SELECT * FROM decks WHERE id = ?")
+      ).get([newDeck.id])) as { filters: unknown };
 
       expect(deckExists).toBeDefined();
       expect(deckExists.filters).toBeNull();
@@ -96,17 +87,12 @@ describe("decksTable", () => {
         },
       } satisfies Omit<SelectDeck, "id">;
 
-      const newDeck = await decksTable.create.mutation({
-        deck: deckInput,
-      });
-
+      const newDeck = await decksTable.create.mutation({ deck: deckInput });
       const updatedName = "updated deck";
 
       const updatedDeck = await decksTable.update.mutation({
         id: newDeck.id,
-        updates: {
-          name: updatedName,
-        },
+        updates: { name: updatedName },
       });
 
       expect(updatedDeck).toMatchObject({
@@ -118,13 +104,6 @@ describe("decksTable", () => {
           tags: expect.arrayContaining(deckInput.filters.tags),
         },
       });
-
-      const deckExists = await testDb.db
-        .prepare("SELECT * FROM decks WHERE id = ?")
-        .get([newDeck.id]);
-
-      expect(deckExists).toBeDefined();
-      expect(deckExists.name).toBe(updatedName);
     });
 
     it("updates the filters field", async () => {
@@ -152,80 +131,49 @@ describe("decksTable", () => {
           tags: expect.arrayContaining(updatedFilters.tags),
         },
       });
-
-      const deckExists = await testDb.db
-        .prepare("SELECT * FROM decks WHERE id = ?")
-        .get([newDeck.id]);
-
-      expect(JSON.parse(deckExists.filters)).toMatchObject(updatedFilters);
     });
 
     it("throws when no fields are provided", async () => {
       const newDeck = await decksTable.create.mutation({
-        deck: {
-          name: "test deck",
-          filters: null,
-        },
+        deck: { name: "test deck", filters: null },
       });
 
-      const updateResult = decksTable.update.mutation({
-        id: newDeck.id,
-        updates: {},
-      });
-
-      await expect(updateResult).rejects.toThrow("No fields to update");
+      await expect(
+        decksTable.update.mutation({ id: newDeck.id, updates: {} })
+      ).rejects.toThrow("No fields to update");
     });
 
     it("throws when the deck does not exist", async () => {
-      const updateResult = decksTable.update.mutation({
-        id: "not-a-real-id",
-        updates: {
-          name: "updated deck",
-        },
-      });
-
-      await expect(updateResult).rejects.toThrow("Deck not found");
+      await expect(
+        decksTable.update.mutation({
+          id: "not-a-real-id",
+          updates: { name: "updated deck" },
+        })
+      ).rejects.toThrow("Deck not found");
     });
   });
 
   describe("delete", () => {
     it("removes the deck", async () => {
-      const deckInput = {
-        name: "test deck",
-        filters: {
-          state: [FlashcardState.LEARNING, FlashcardState.REVIEW],
-          types: ["ism", "fi'l"],
-          tags: ["foo"],
-        },
-      } satisfies Omit<SelectDeck, "id">;
-
       const newDeck = await decksTable.create.mutation({
-        deck: deckInput,
+        deck: { name: "test deck", filters: null },
       });
 
-      const deleteDeckResult = decksTable.delete.mutation({
-        id: newDeck.id,
-      });
+      await expect(
+        decksTable.delete.mutation({ id: newDeck.id })
+      ).resolves.toMatchObject({ success: true });
 
-      await expect(deleteDeckResult).resolves.toMatchObject({
-        success: true,
-      });
-
-      const deckExists = await testDb.db
-        .prepare("SELECT * FROM decks WHERE id = ?")
-        .get([newDeck.id]);
+      const deckExists = await (
+        await testDb.db.prepare("SELECT * FROM decks WHERE id = ?")
+      ).get([newDeck.id]);
 
       expect(deckExists).toBeUndefined();
     });
 
     it("does not throw when deleting a deck that doesn't exist", async () => {
-      const deleteDeckResult = decksTable.delete.mutation({
-        id: "not-a-real-id",
-      });
-
-      await expect(deleteDeckResult).resolves.toMatchObject({
-        success: true,
-      });
+      await expect(
+        decksTable.delete.mutation({ id: "not-a-real-id" })
+      ).resolves.toMatchObject({ success: true });
     });
   });
 
@@ -237,15 +185,11 @@ describe("decksTable", () => {
         const ismEntry = await insertDictionaryEntry(testDb, { type: "ism" });
         await insertFlashcard(testDb, { dictionary_entry_id: ismEntry.id });
 
-        const filEntry = await insertDictionaryEntry(testDb, {
-          type: "fi'l",
-        });
+        const filEntry = await insertDictionaryEntry(testDb, { type: "fi'l" });
         await insertFlashcard(testDb, { dictionary_entry_id: filEntry.id });
 
         const decks = await decksTable.list.query({});
-        const result = decks.find((d) => d.id === deck.id);
-
-        expect(result?.total_hits).toBe(1);
+        expect(decks.find((d) => d.id === deck.id)?.total_hits).toBe(1);
       });
 
       it("defaults to counting all word types when types is unset", async () => {
@@ -257,9 +201,7 @@ describe("decksTable", () => {
         }
 
         const decks = await decksTable.list.query({});
-        const result = decks.find((d) => d.id === deck.id);
-
-        expect(result?.total_hits).toBe(4);
+        expect(decks.find((d) => d.id === deck.id)?.total_hits).toBe(4);
       });
     });
 
@@ -273,9 +215,7 @@ describe("decksTable", () => {
         await insertFlashcard(testDb, { state: FlashcardState.REVIEW });
 
         const decks = await decksTable.list.query({});
-        const result = decks.find((d) => d.id === deck.id);
-
-        expect(result?.total_hits).toBe(1);
+        expect(decks.find((d) => d.id === deck.id)?.total_hits).toBe(1);
       });
 
       it("defaults to counting all states when state is unset", async () => {
@@ -287,9 +227,7 @@ describe("decksTable", () => {
         await insertFlashcard(testDb, { state: FlashcardState.RE_LEARNING });
 
         const decks = await decksTable.list.query({});
-        const result = decks.find((d) => d.id === deck.id);
-
-        expect(result?.total_hits).toBe(4);
+        expect(decks.find((d) => d.id === deck.id)?.total_hits).toBe(4);
       });
     });
 
@@ -297,39 +235,27 @@ describe("decksTable", () => {
       it("counts only flashcards whose dictionary entry has a specified tag", async () => {
         const deck = await insertDeck(testDb, { filters: { tags: ["foo"] } });
 
-        const fooEntry = await insertDictionaryEntry(testDb, {
-          tags: ["foo"],
-        });
+        const fooEntry = await insertDictionaryEntry(testDb, { tags: ["foo"] });
         await insertFlashcard(testDb, { dictionary_entry_id: fooEntry.id });
 
-        const barEntry = await insertDictionaryEntry(testDb, {
-          tags: ["bar"],
-        });
+        const barEntry = await insertDictionaryEntry(testDb, { tags: ["bar"] });
         await insertFlashcard(testDb, { dictionary_entry_id: barEntry.id });
 
         const decks = await decksTable.list.query({});
-        const result = decks.find((d) => d.id === deck.id);
-
-        expect(result?.total_hits).toBe(1);
+        expect(decks.find((d) => d.id === deck.id)?.total_hits).toBe(1);
       });
 
       it("does not restrict by tag when tags is unset", async () => {
         const deck = await insertDeck(testDb, { filters: null });
 
-        const fooEntry = await insertDictionaryEntry(testDb, {
-          tags: ["foo"],
-        });
+        const fooEntry = await insertDictionaryEntry(testDb, { tags: ["foo"] });
         await insertFlashcard(testDb, { dictionary_entry_id: fooEntry.id });
 
-        const noTagEntry = await insertDictionaryEntry(testDb, {
-          tags: null,
-        });
+        const noTagEntry = await insertDictionaryEntry(testDb, { tags: null });
         await insertFlashcard(testDb, { dictionary_entry_id: noTagEntry.id });
 
         const decks = await decksTable.list.query({});
-        const result = decks.find((d) => d.id === deck.id);
-
-        expect(result?.total_hits).toBe(2);
+        expect(decks.find((d) => d.id === deck.id)?.total_hits).toBe(2);
       });
     });
 
@@ -341,9 +267,7 @@ describe("decksTable", () => {
         await insertFlashcard(testDb, { direction: "reverse" });
 
         const decks = await decksTable.list.query({ show_reverse: false });
-        const result = decks.find((d) => d.id === deck.id);
-
-        expect(result?.total_hits).toBe(1);
+        expect(decks.find((d) => d.id === deck.id)?.total_hits).toBe(1);
       });
 
       it("includes reverse-direction flashcards when show_reverse is true", async () => {
@@ -353,9 +277,7 @@ describe("decksTable", () => {
         await insertFlashcard(testDb, { direction: "reverse" });
 
         const decks = await decksTable.list.query({ show_reverse: true });
-        const result = decks.find((d) => d.id === deck.id);
-
-        expect(result?.total_hits).toBe(2);
+        expect(decks.find((d) => d.id === deck.id)?.total_hits).toBe(2);
       });
     });
 
@@ -425,9 +347,7 @@ describe("decksTable", () => {
         });
 
         const decks = await decksTable.list.query({ show_reverse: false });
-        const result = decks.find((d) => d.id === deck.id);
-
-        expect(result?.total_hits).toBe(1);
+        expect(decks.find((d) => d.id === deck.id)?.total_hits).toBe(1);
       });
     });
 
@@ -462,10 +382,8 @@ describe("decksTable", () => {
 
     describe("due-date windowing", () => {
       it("excludes not-yet-due flashcards from to_review/to_review_backlog, but total_hits still counts them", async () => {
-        // total_hits is `COUNT(DISTINCT f.id)` with no due-date condition at
-        // all -- only to_review/to_review_backlog filter by due date. Pinning
-        // this asymmetry down since it's easy to assume total_hits means
-        // "total due" when it actually means "total matching the filters".
+        // total_hits is COUNT(DISTINCT f.id) with no due-date condition --
+        // only to_review/to_review_backlog filter by due date.
         const deck = await insertDeck(testDb, { filters: null });
 
         const futureDate = new Date(Date.now() + 1000 * 60 * 60 * 24 * 2);
@@ -486,14 +404,14 @@ describe("decksTable", () => {
         const deck = await insertDeck(testDb, { filters: null });
         const backlogThresholdDays = 7;
 
-        // Due 1 day ago: within the backlog threshold -> to_review.
+        // Due 1 day ago: within threshold -> to_review.
         const recentlyDueDate = new Date(Date.now() - 1000 * 60 * 60 * 24 * 1);
         await insertFlashcard(testDb, {
           due: recentlyDueDate.toISOString(),
           due_timestamp_ms: recentlyDueDate.getTime(),
         });
 
-        // Due 10 days ago: past the backlog threshold -> to_review_backlog.
+        // Due 10 days ago: past threshold -> to_review_backlog.
         const oldDueDate = new Date(Date.now() - 1000 * 60 * 60 * 24 * 10);
         await insertFlashcard(testDb, {
           due: oldDueDate.toISOString(),
