+++ packages/db-operations/src/operations/dictionary-entries.test.ts
@@ -1,25 +1,18 @@
 import type { SelectDictionaryEntry } from "@bahar/drizzle-user-db-schemas";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { createTestDb, type TestDb } from "../test/create-test-db";
 import { insertDictionaryEntry, insertFlashcard } from "../test/factories";
-
-const dbRef = vi.hoisted(() => ({ current: undefined as TestDb | undefined }));
-
-vi.mock("..", async (importOriginal) => ({
-  ...(await importOriginal()),
-  ensureDb: vi.fn(async () => dbRef.current?.db),
-  getDb: vi.fn(() => dbRef.current?.db),
-  getDrizzleDb: vi.fn(() => dbRef.current?.drizzleDb),
-}));
-
-const { dictionaryEntriesTable } = await import("./dictionary-entries");
+import { makeDictionaryEntriesTable } from "./dictionary-entries";
 
 describe("dictionaryEntriesTable", () => {
   let testDb: TestDb;
+  let dictionaryEntriesTable: ReturnType<typeof makeDictionaryEntriesTable>;
 
   beforeEach(async () => {
     testDb = await createTestDb();
-    dbRef.current = testDb;
+    dictionaryEntriesTable = makeDictionaryEntriesTable({
+      getDb: async () => testDb.drizzleDb,
+    });
   });
 
   afterEach(async () => {
@@ -43,9 +36,9 @@ describe("dictionaryEntriesTable", () => {
     });
 
     it("throws when the entry does not exist", async () => {
-      const entryResult = dictionaryEntriesTable.entry.query("not-a-real-id");
-
-      await expect(entryResult).rejects.toThrow("Dictionary entry not found");
+      await expect(
+        dictionaryEntriesTable.entry.query("not-a-real-id")
+      ).rejects.toThrow("Dictionary entry not found");
     });
   });
 
@@ -90,21 +83,11 @@ describe("dictionaryEntriesTable", () => {
         type: "ism",
         tags: ["foo"],
       });
-
-      const entryExists = await testDb.db
-        .prepare("SELECT * FROM dictionary_entries WHERE id = ?")
-        .get([newEntry.id]);
-
-      expect(entryExists).toBeDefined();
     });
 
     it("stores unset optional JSON fields as SQL NULL, not the string 'null'", async () => {
       const newEntry = await dictionaryEntriesTable.addWord.mutation({
-        word: {
-          word: "كتاب",
-          translation: "book",
-          type: "ism",
-        },
+        word: { word: "كتاب", translation: "book", type: "ism" },
       });
 
       expect(newEntry.root).toBeNull();
@@ -113,15 +96,12 @@ describe("dictionaryEntriesTable", () => {
       expect(newEntry.examples).toBeNull();
       expect(newEntry.morphology).toBeNull();
 
-      const row = await testDb.db
-        .prepare("SELECT * FROM dictionary_entries WHERE id = ?")
-        .get([newEntry.id]);
+      const row = (await (
+        await testDb.db.prepare("SELECT * FROM dictionary_entries WHERE id = ?")
+      ).get([newEntry.id])) as Record<string, unknown>;
 
       expect(row.root).toBeNull();
       expect(row.tags).toBeNull();
-      expect(row.antonyms).toBeNull();
-      expect(row.examples).toBeNull();
-      expect(row.morphology).toBeNull();
     });
   });
 
@@ -134,16 +114,14 @@ describe("dictionaryEntriesTable", () => {
         tags: ["foo"],
       });
 
-      const updatedWord = "قلم";
-
       const updatedEntry = await dictionaryEntriesTable.editWord.mutation({
         id: entry.id,
-        updates: { word: updatedWord },
+        updates: { word: "قلم" },
       });
 
       expect(updatedEntry).toMatchObject({
         id: entry.id,
-        word: updatedWord,
+        word: "قلم",
         translation: "book",
         type: "ism",
         tags: ["foo"],
@@ -151,12 +129,6 @@ describe("dictionaryEntriesTable", () => {
       expect(updatedEntry.updated_at_timestamp_ms).toBeGreaterThanOrEqual(
         entry.updated_at_timestamp_ms ?? 0
       );
-
-      const entryExists = await testDb.db
-        .prepare("SELECT * FROM dictionary_entries WHERE id = ?")
-        .get([entry.id]);
-
-      expect(entryExists.word).toBe(updatedWord);
     });
 
     it("updates every other field", async () => {
@@ -188,27 +160,12 @@ describe("dictionaryEntriesTable", () => {
       });
 
       expect(updatedEntry).toMatchObject(updates);
-
-      const row = await testDb.db
-        .prepare("SELECT * FROM dictionary_entries WHERE id = ?")
-        .get([entry.id]);
-
-      expect(row.translation).toBe(updates.translation);
-      expect(row.definition).toBe(updates.definition);
-      expect(row.type).toBe(updates.type);
-      expect(JSON.parse(row.root)).toEqual(updates.root);
-      expect(JSON.parse(row.tags)).toEqual(updates.tags);
-      expect(JSON.parse(row.antonyms)).toEqual(updates.antonyms);
-      expect(JSON.parse(row.examples)).toEqual(updates.examples);
-      expect(JSON.parse(row.morphology)).toEqual(updates.morphology);
     });
 
     it("does not throw when no fields are provided, and only bumps updated_at", async () => {
-      // Unlike decksTable.update/flashcardsTable.update/settingsTable.update,
-      // editWord has no "No fields to update" guard -- it always pushes
-      // updated_at/updated_at_timestamp_ms onto setClauses unconditionally,
-      // so that array is never actually empty. Pinning down the real
-      // (inconsistent-with-its-siblings) behavior rather than the assumed one.
+      // Unlike the other tables' update ops, editWord has no "No fields to
+      // update" guard -- it always pushes updated_at, so the set is never
+      // empty. Pinning down the real (inconsistent-with-siblings) behavior.
       const entry = await insertDictionaryEntry(testDb, {
         word: "كتاب",
         translation: "book",
@@ -219,22 +176,19 @@ describe("dictionaryEntriesTable", () => {
         updates: {},
       });
 
-      expect(updatedEntry).toMatchObject({
-        word: "كتاب",
-        translation: "book",
-      });
+      expect(updatedEntry).toMatchObject({ word: "كتاب", translation: "book" });
       expect(updatedEntry.updated_at_timestamp_ms).toBeGreaterThanOrEqual(
         entry.updated_at_timestamp_ms ?? 0
       );
     });
 
     it("throws when the entry does not exist", async () => {
-      const editResult = dictionaryEntriesTable.editWord.mutation({
-        id: "not-a-real-id",
-        updates: { word: "قلم" },
-      });
-
-      await expect(editResult).rejects.toThrow("Dictionary entry not found");
+      await expect(
+        dictionaryEntriesTable.editWord.mutation({
+          id: "not-a-real-id",
+          updates: { word: "قلم" },
+        })
+      ).rejects.toThrow("Dictionary entry not found");
     });
   });
 
@@ -245,26 +199,24 @@ describe("dictionaryEntriesTable", () => {
         translation: "book",
       });
 
-      const deleteResult = dictionaryEntriesTable.delete.mutation({
-        id: entry.id,
-      });
-
-      await expect(deleteResult).resolves.toMatchObject({
+      await expect(
+        dictionaryEntriesTable.delete.mutation({ id: entry.id })
+      ).resolves.toMatchObject({
         id: entry.id,
         word: "كتاب",
         translation: "book",
       });
 
-      const entryExists = await testDb.db
-        .prepare("SELECT * FROM dictionary_entries WHERE id = ?")
-        .get([entry.id]);
+      const exists = await (
+        await testDb.db.prepare("SELECT * FROM dictionary_entries WHERE id = ?")
+      ).get([entry.id]);
 
-      expect(entryExists).toBeUndefined();
+      expect(exists).toBeUndefined();
     });
 
     it("also deletes flashcards linked to the entry", async () => {
-      // sync-wasm doesn't support ON DELETE CASCADE, so dictionaryEntriesTable.delete
-      // does this manually -- pin it down so the drizzle refactor doesn't drop it.
+      // sync-wasm has no ON DELETE CASCADE, so delete does this manually --
+      // pin it down so the shared impl keeps it.
       const entry = await insertDictionaryEntry(testDb);
       const flashcard = await insertFlashcard(testDb, {
         dictionary_entry_id: entry.id,
@@ -272,22 +224,17 @@ describe("dictionaryEntriesTable", () => {
 
       await dictionaryEntriesTable.delete.mutation({ id: entry.id });
 
-      const flashcardExists = await testDb.db
-        .prepare("SELECT * FROM flashcards WHERE id = ?")
-        .get([flashcard.id]);
+      const flashcardExists = await (
+        await testDb.db.prepare("SELECT * FROM flashcards WHERE id = ?")
+      ).get([flashcard.id]);
 
       expect(flashcardExists).toBeUndefined();
     });
 
     it("throws when the entry does not exist", async () => {
-      // Differs from decksTable.delete, which does NOT throw for a missing id --
-      // pinning this down since the two tables currently behave differently
-      // for the same kind of operation.
-      const deleteResult = dictionaryEntriesTable.delete.mutation({
-        id: "not-a-real-id",
-      });
-
-      await expect(deleteResult).rejects.toThrow("Dictionary entry not found");
+      await expect(
+        dictionaryEntriesTable.delete.mutation({ id: "not-a-real-id" })
+      ).rejects.toThrow("Dictionary entry not found");
     });
   });
 
@@ -296,15 +243,11 @@ describe("dictionaryEntriesTable", () => {
       await insertDictionaryEntry(testDb, { updated_at_timestamp_ms: 1000 });
       await insertDictionaryEntry(testDb, { updated_at_timestamp_ms: 2000 });
 
-      const result = await dictionaryEntriesTable.maxUpdatedAt.query();
-
-      expect(result).toBe(2000);
+      expect(await dictionaryEntriesTable.maxUpdatedAt.query()).toBe(2000);
     });
 
     it("returns null when there are no entries", async () => {
-      const result = await dictionaryEntriesTable.maxUpdatedAt.query();
-
-      expect(result).toBeNull();
+      expect(await dictionaryEntriesTable.maxUpdatedAt.query()).toBeNull();
     });
   });
 });
