+++ packages/db-operations/src/operations/settings.test.ts
@@ -1,24 +1,15 @@
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import { afterEach, beforeEach, describe, expect, it } from "vitest";
 import { createTestDb, type TestDb } from "../test/create-test-db";
 import { insertSettings } from "../test/factories";
-
-const dbRef = vi.hoisted(() => ({ current: undefined as TestDb | undefined }));
-
-vi.mock("..", async (importOriginal) => ({
-  ...(await importOriginal()),
-  ensureDb: vi.fn(async () => dbRef.current?.db),
-  getDb: vi.fn(() => dbRef.current?.db),
-  getDrizzleDb: vi.fn(() => dbRef.current?.drizzleDb),
-}));
-
-const { settingsTable } = await import("./settings");
+import { makeSettingsTable } from "./settings";
 
 describe("settingsTable", () => {
   let testDb: TestDb;
+  let settingsTable: ReturnType<typeof makeSettingsTable>;
 
   beforeEach(async () => {
     testDb = await createTestDb();
-    dbRef.current = testDb;
+    settingsTable = makeSettingsTable({ getDb: async () => testDb.drizzleDb });
   });
 
   afterEach(async () => {
@@ -27,14 +18,16 @@ describe("settingsTable", () => {
 
   describe("getSettings", () => {
     it("creates and returns default settings when none exist yet", async () => {
-      const settings = await settingsTable.getSettings.query();
+      const result = await settingsTable.getSettings.query();
 
-      expect(settings).toEqual({
+      expect(result).toEqual({
         show_antonyms_in_flashcard: "hidden",
         show_reverse_flashcards: false,
       });
 
-      const row = await testDb.db.prepare("SELECT * FROM settings").get();
+      const row = await (
+        await testDb.db.prepare("SELECT * FROM settings")
+      ).get();
 
       expect(row).toBeDefined();
     });
@@ -45,9 +38,9 @@ describe("settingsTable", () => {
         show_reverse_flashcards: true,
       });
 
-      const settings = await settingsTable.getSettings.query();
+      const result = await settingsTable.getSettings.query();
 
-      expect(settings).toEqual({
+      expect(result).toEqual({
         show_antonyms_in_flashcard: "hint",
         show_reverse_flashcards: true,
       });
@@ -70,7 +63,12 @@ describe("settingsTable", () => {
         show_reverse_flashcards: true,
       });
 
-      const row = await testDb.db.prepare("SELECT * FROM settings").get();
+      const row = (await (
+        await testDb.db.prepare("SELECT * FROM settings")
+      ).get()) as {
+        show_reverse_flashcards: number;
+        show_antonyms_in_flashcard: string;
+      };
 
       expect(row.show_reverse_flashcards).toBe(1);
       expect(row.show_antonyms_in_flashcard).toBe("hidden");
@@ -90,29 +88,22 @@ describe("settingsTable", () => {
         show_antonyms_in_flashcard: "hint",
         show_reverse_flashcards: false,
       });
-
-      const row = await testDb.db.prepare("SELECT * FROM settings").get();
-
-      expect(row.show_antonyms_in_flashcard).toBe("hint");
-      expect(row.show_reverse_flashcards).toBe(0);
     });
 
     it("throws when no fields are provided", async () => {
       await insertSettings(testDb);
 
-      const updateResult = settingsTable.update.mutation({ updates: {} });
-
-      await expect(updateResult).rejects.toThrow("No fields to update");
+      await expect(
+        settingsTable.update.mutation({ updates: {} })
+      ).rejects.toThrow("No fields to update");
     });
-  });
 
-  describe("update called before any settings row exists", () => {
-    it("throws a controlled error instead of crashing", async () => {
-      const updateResult = settingsTable.update.mutation({
-        updates: { show_reverse_flashcards: true },
-      });
-
-      await expect(updateResult).rejects.toThrow("Settings not found");
+    it("throws a controlled error when no settings row exists", async () => {
+      await expect(
+        settingsTable.update.mutation({
+          updates: { show_reverse_flashcards: true },
+        })
+      ).rejects.toThrow("Settings not found");
     });
   });
 });
